+++ src/club_existing/definitions_patches.d.ts
@@ -6,7 +6,3 @@ type NotificationSetting = any;
 type HSVColor = { H: number, S: number, V: number };
 
 type VibratorMode = "Off" | "Low" | "Medium" | "High" | "Maximum" | "Random" | "Escalate" | "Tease" | "Deny" | "Edge";
-
-interface ItemProperties {
-	AllowActivityOn_BCXBackup?: AssetGroupName[];
-}